requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.3
geographiclib==1.50
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
whitenoise==5.2.0
python-decouple==3.4
django-filter==2.4.0
django-crispy==forms-1.12.0
geocoder==1.38.1
